fix: Change Inventory module name to Store
