--- src/modules/Inventory/composables/Movement/useIndex.ts
@@ -1,89 +1,0 @@
-import { reactive, onMounted } from "vue"
-import { onBeforeRouteUpdate } from "vue-router"
-import useTableGrid from "@/composables/useTableGrid"
-import useHttp from "@/composables/useHttp"
-import MovementService from "../../services/Movement"
-
-type Params =  string | string[][] | Record<string, string> | URLSearchParams | undefined
-
-export default () => {
-  const data = reactive({
-    rows: [],
-    links: [],
-    search: "",
-    sort: "",
-    direction: ""
-  })
-
-  const {  
-    errors,
-
-    getError     
-  } = useHttp()
-
-  const {
-    route,
-    router,
-
-    setSearch,
-    setSort, 
-  } = useTableGrid(data, "/movements")
-
-  const getMovements = (routeQuery: string) => {
-    return MovementService.getMovements(routeQuery)
-      .then((response) => {
-        errors.value = {}
-        data.rows = response.data.rows.data
-        data.links = response.data.rows.links
-        data.search = response.data.search
-        data.sort = response.data.sort
-        data.direction = response.data.direction      
-      })
-      .catch((error) => {
-        console.log(error)
-      })
-  }
-
-  const deleteRow = (rowId?: string) => {
-    if (rowId === undefined)
-      return
-    else if (confirm(`¿Estás seguro que desea eliminar el registro ${rowId}?`)) {    
-      return MovementService.deleteMovement(rowId)
-        .then((response) => {
-          errors.value = {}
-          getMovements(
-            new URLSearchParams(route.query as Params).toString()
-          )
-        })
-        .catch((err) => {                
-          console.log( err.response.data )
-          errors.value = getError(err)
-        })
-    }
-  }
-
-  onBeforeRouteUpdate(async (to, from) => {      
-    if (to.query !== from.query) {        
-      await getMovements(
-        new URLSearchParams(to.query as Params).toString()
-      )
-    }
-  })
-
-  onMounted(() => {
-    getMovements(
-      new URLSearchParams(route.query as Params).toString()
-    )
-  })
-
-  return {
-    errors,
-    data,
-    router,
-
-    deleteRow,
-    setSearch,
-    setSort
-  }
-}
-
