--- src/modules/Inventory/types/Role.ts
@@ -1,4 +1,0 @@
-export default interface Role {    
-  id: number;
-  name: string;
-}
