--- src/modules/Store/types/Errors.ts
@@ -1,6 +1,0 @@
-export default interface Errors {    
-  name?: [];
-  email?: [];
-  password?: [];
-  role_id?: [];
-}
