--- src/modules/Store/types/User.ts
@@ -1,7 +1,0 @@
-export default interface User {
-  id?: string;
-  name: string | null;
-  email: string | null;
-  password: string | null;
-  role_id: string | null;
-}
