+++ src/router/index.ts
@@ -6,7 +6,6 @@ import middlewarePipeline from "@/router/middlewarePipeline"
 import AuthRoutes from "@/modules/Auth/routes"
 import AuthorizationRoutes from "@/modules/Authorization/routes"
 import ArticleRoutes from "@/modules/Article/routes"
-import InventoryRoutes from "@/modules/Inventory/routes"
 import ProductRoutes from "@/modules/Product/routes"
 import StoreRoutes from "@/modules/Store/routes"
 import UserRoutes from "@/modules/User/routes"
@@ -17,7 +16,6 @@ const routes: Array<RouteRecordRaw> = [
   ...AuthRoutes.map(route => route),
   ...AuthorizationRoutes.map(route => route),
   ...ArticleRoutes.map(route => route),
-  ...InventoryRoutes.map(route => route),
   ...ProductRoutes.map(route => route),
   ...StoreRoutes.map(route => route),
   ...UserRoutes.map(route => route)
